api/src/requirements.txt
Django==3.2.9
gunicorn==20.1.0
pytz==2021.3
sqlparse==0.4.2
<<<<<<< HEAD
whitenoise==5.3.0
dj-database-url==0.5.0
=======
<<<<<<< Updated upstream
# whitenoise==5.2.0
dj-database-url==0.5.0
=======
whitenoise==5.3.0
dj-database-url==0.5.0
djangorestframework==3.12.4
django-filter==21.1
django-cors-headers==3.10.0
psycopg2==2.9.1
>>>>>>> Stashed changes
>>>>>>> e283125... updated requirements.txt for docker, removed merge conflicts
